chore: move ts rule's test.ts to js rule's dir
