+++ packages/eslint-plugin-js/rules/brace-style/ts-brace-style.test.ts
