+++ packages/eslint-plugin-js/rules/key-spacing/ts-key-spacing.test.ts
