+++ packages/eslint-plugin-js/rules/no-extra-parens/ts-no-extra-parens.test.ts
