+++ packages/eslint-plugin-js/rules/no-extra-semi/ts-no-extra-semi.test.ts
