+++ packages/eslint-plugin-js/rules/object-curly-spacing/ts-object-curly-spacing.test.ts
