+++ packages/eslint-plugin-js/rules/object-property-newline/ts-object-property-newline.test.ts
