+++ packages/eslint-plugin-js/rules/quote-props/ts-quote-props.test.ts
