+++ packages/eslint-plugin-js/rules/semi/ts-semi.test.ts
