+++ packages/eslint-plugin-js/rules/space-before-function-paren/ts-space-before-function-paren.test.ts
