+++ packages/eslint-plugin-js/rules/space-infix-ops/ts-space-infix-ops.test.ts
